Fixed the db context problem

+++ JurassicParkCore/DataSchemas/AnimalType.cs
@@ -4,7 +4,7 @@ namespace JurassicParkCore.DataSchemas;
 
 public class AnimalType : IKeyedDataType
 {
-    [Key] public int Id { get; set; }
+    [Key] public long Id { get; set; }
     [MaxLength(20)] public required string Name { get; set; }
     public required AnimalEatingHabit EatingHabit { get; set; }
     public required decimal Price { get; set; }
