+++ JurassicParkCore/DataSchemas/IKeyedDataType.cs
@@ -2,5 +2,5 @@ namespace JurassicParkCore.DataSchemas;
 
 public interface IKeyedDataType
 {
-    public int Id { get; }
+    public long Id { get; }
 }
